Started to orgnanizae hero states and split animation info into own file
